# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Caramurú Field Engineering API",
    "description": "API da plataforma de engenharia de campo da Caramurú Construções. Multi-tenant, obras (subestações e LTs) com PostGIS e ativos com QR Code.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/safety/aprs": {
      "post": {
        "tags": [
          "safety"
        ],
        "summary": "Create Apr",
        "operationId": "create_apr_api_v1_safety_aprs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SafetyInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafetyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "safety"
        ],
        "summary": "List Aprs",
        "operationId": "list_aprs_api_v1_safety_aprs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SafetyResponse"
                  },
                  "title": "Response List Aprs Api V1 Safety Aprs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/safety/aprs/{record_id}": {
      "get": {
        "tags": [
          "safety"
        ],
        "summary": "Get Apr",
        "operationId": "get_apr_api_v1_safety_aprs__record_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Record Id"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafetyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/safety/aprs/{record_id}/signatures": {
      "post": {
        "tags": [
          "safety"
        ],
        "summary": "Sign Apr",
        "operationId": "sign_apr_api_v1_safety_aprs__record_id__signatures_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Record Id"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignatureInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafetyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/safety/aprs/{record_id}/release": {
      "post": {
        "tags": [
          "safety"
        ],
        "summary": "Release Apr",
        "operationId": "release_apr_api_v1_safety_aprs__record_id__release_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Record Id"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafetyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/safety/permits": {
      "post": {
        "tags": [
          "safety"
        ],
        "summary": "Create Permit",
        "operationId": "create_permit_api_v1_safety_permits_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermitInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/safety/permits/{record_id}": {
      "get": {
        "tags": [
          "safety"
        ],
        "summary": "Get Permit",
        "operationId": "get_permit_api_v1_safety_permits__record_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Record Id"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/safety/permits/{record_id}/signatures": {
      "post": {
        "tags": [
          "safety"
        ],
        "summary": "Sign Permit",
        "operationId": "sign_permit_api_v1_safety_permits__record_id__signatures_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Record Id"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignatureInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/safety/permits/{record_id}/release": {
      "post": {
        "tags": [
          "safety"
        ],
        "summary": "Release Permit",
        "operationId": "release_permit_api_v1_safety_permits__record_id__release_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Record Id"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PermitInput": {
        "properties": {
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "observation": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 1,
            "title": "Observation"
          },
          "voltage_kv": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Voltage Kv"
          },
          "work_height_m": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Height M"
          },
          "risk_level": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "CRITICAL"
            ],
            "title": "Risk Level",
            "default": "MEDIUM"
          },
          "epi_checklist": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Epi Checklist"
          },
          "epc_checklist": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Epc Checklist"
          },
          "hazards": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Hazards"
          },
          "valid_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          },
          "valid_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Until"
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "activity": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 1,
            "title": "Activity"
          },
          "apr_id": {
            "type": "string",
            "format": "uuid",
            "title": "Apr Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project_id",
          "observation",
          "title",
          "activity",
          "apr_id"
        ],
        "title": "PermitInput"
      },
...
      "SafetyInput": {
        "properties": {
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "observation": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 1,
            "title": "Observation"
          },
          "voltage_kv": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Voltage Kv"
          },
          "work_height_m": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Height M"
          },
          "risk_level": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "CRITICAL"
            ],
            "title": "Risk Level",
            "default": "MEDIUM"
          },
          "epi_checklist": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Epi Checklist"
          },
          "epc_checklist": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Epc Checklist"
          },
          "hazards": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Hazards"
          },
          "valid_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          },
          "valid_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Until"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project_id",
          "observation"
        ],
        "title": "SafetyInput"
      },
      "SafetyResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "requires_nr10": {
            "type": "boolean",
            "title": "Requires Nr10"
          },
          "requires_nr35": {
            "type": "boolean",
            "title": "Requires Nr35"
          },
          "voltage_kv": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voltage Kv"
          },
          "work_height_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Height M"
          },
          "risk_level": {
            "type": "string",
            "title": "Risk Level"
          },
          "epi_checklist": {
            "additionalProperties": true,
            "type": "object",
            "title": "Epi Checklist"
          },
          "epc_checklist": {
            "additionalProperties": true,
            "type": "object",
            "title": "Epc Checklist"
          },
          "hazards": {
            "items": {},
            "type": "array",
            "title": "Hazards"
          },
          "tst_signature": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tst Signature"
          },
          "encarregado_signature": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Encarregado Signature"
          },
          "valid_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          },
          "valid_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Until"
          },
          "released_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Released At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "status",
          "requires_nr10",
          "requires_nr35",
          "voltage_kv",
          "work_height_m",
          "risk_level",
          "epi_checklist",
          "epc_checklist",
          "hazards",
          "tst_signature",
          "encarregado_signature",
          "valid_from",
          "valid_until",
          "released_at"
        ],
        "title": "SafetyResponse"
      },
      "SignatureInput": {
        "properties": {
          "signer_role": {
            "type": "string",
            "enum": [
              "TST",
              "ENCARREGADO"
            ],
            "title": "Signer Role"
          },
          "signature_svg": {
            "type": "string",
            "maxLength": 100000,
            "minLength": 20,
            "title": "Signature Svg"
          },
          "sha256": {
            "type": "string",
            "pattern": "^[0-9a-fA-F]{64}$",
            "title": "Sha256"
          },
          "signed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Signed At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "signer_role",
          "signature_svg",
          "sha256",
          "signed_at"
        ],
        "title": "SignatureInput"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}